Added Threshold policies : hard will be hard to trigger normal will trigger normaly light will trigger often Theses trigger policies are associated with priorities low medium and high. A priority high will have more chances to happen.
The threshold is set at the beggining of the simulation whereas the priorities are set for each component modification (add/remove)

When a component a component add or remove is called, the probability associated to the threshold and priority is compared to a random value.
If the random value is lesser, the modification is triggered.

With this update, each simulation will be different because policy and threshold probabilities are taken according to ranges.

An exception is made with (setSignal) methods because they are emergency modifications and they must be triggered.
The probailistic choice is made for now due to the easy of results but chance be changed to something else

Some incoming improvements :
Battery >100
Frequency of threshold
start/stop reconfiguration